examples/requirements.examples.txt
python-dotenv
openai
anthropic
google-genai
